requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autoflake==2.3.1
billiard==4.2.1
black==25.1.0
build==1.2.2.post1
celery==5.5.2
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.8.0
cron-descriptor==1.4.5
decorator==5.2.1
distlib==0.3.9
Django==5.1.7
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docformatter==1.7.7
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
exceptiongroup==1.2.2
executing==2.2.0
factory_boy==3.3.3
Faker==37.3.0
filelock==3.18.0
flake8==7.2.0
flower==2.0.1
freezegun==1.5.1
gunicorn==23.0.0
humanize==4.12.3
identify==2.6.12
importlib_metadata==8.7.0
inflection==0.5.1
ipython==8.34.0
isort==6.0.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.8
pre_commit==4.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
pure_eval==0.2.3
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.23.1
ruff==0.12.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
untokenize==0.1.1
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
zipp==3.23.0
